:hammer: add all
