--- src/hooks/useAuth.js
@@ -1,29 +1,0 @@
-// src/app/hooks/useAuth.js
-"use client";
-
-import { useState, useEffect } from "react";
-
-export const useAuth = () => {
-    const [user, setUser] = useState(null);
-    const [isLoading, setIsLoading] = useState(true);
-
-    useEffect(() => {
-        const fetchUser = async () => {
-            try {
-                const res = await fetch("/api/user", {
-                    credentials: "include",
-                });
-                const data = await res.json();
-                setUser(data.user);
-            } catch (error) {
-                console.error("Error fetching user:", error);
-            } finally {
-                setIsLoading(false);
-            }
-        };
-
-        fetchUser();
-    }, []);
-
-    return { user, isLoading };
-};
